- Fixed an integer division affecting the layer canvas aspect ratio in arrange layers dialog. - Fixed initialization of the curve line style in the global settings.
git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@1619 5a6a7de5-fb12-0410-b871-c33778c25c60

+++ qtiplot/src/plot2D/dialogs/LayerDialog.cpp
@@ -487,7 +487,7 @@ void LayerDialog::updateSizes(int unit)
 	if (!multi_layer)
 		return;
 
-	aspect_ratio = multi_layer->layerCanvasSize().width()/multi_layer->layerCanvasSize().height();
+	aspect_ratio = (double)multi_layer->layerCanvasSize().width()/(double)multi_layer->layerCanvasSize().height();
 
 	boxCanvasWidth->setValue(convertFromPixels(multi_layer->layerCanvasSize().width(), (FrameWidget::Unit)unit, 0));
 	boxCanvasHeight->setValue(convertFromPixels(multi_layer->layerCanvasSize().height(), (FrameWidget::Unit)unit, 1));
